crear cursos con roles

+++ app/Http/Controllers/CursosController.php
@@ -6,6 +6,7 @@ use App\Models\Categoria;
 use App\Models\Curso;
 use App\Models\Leccion;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CursosController extends Controller
 {
@@ -18,7 +19,10 @@ class CursosController extends Controller
     function lecciones($id)
     {
         $cursos = Curso::findOrFail($id);
+        $this->authorize('editarLecciones', $cursos);
+
         $lecciones = $cursos->lecciones()->orderBy('orden', 'asc')->get();
+        
 
         return view('pages.cursos.leccionesCurso', compact('cursos', 'lecciones'));
     }
@@ -26,12 +30,15 @@ class CursosController extends Controller
     {
         $cursos = Curso::orderBy('id', 'asc')->paginate(10);
         $categories = Categoria::all();
+        
         return view('pages.cursos.table', compact('cursos', 'categories'));
     }
     function index($id)
     {
         $curso = Curso::with(['lecciones'])->findOrFail($id);
-        return view('pages.cursos.index', compact('curso'));
+        $profesor = auth()->user();
+        
+        return view('pages.cursos.index', compact('curso','profesor'));
     }
     public function filterByCategory($id)
     {
@@ -45,7 +52,9 @@ class CursosController extends Controller
      public function destroy($id)
     {   
         $cursos = Curso::find($id);
+        $this->authorize('editarLecciones', $cursos);
         $cursos->delete();
+       
         return redirect()->route('home');
     }
 }
